requirements.txt
pyserial==3.5
redis==4.3.4
python-dotenv==0.21.0
pydantic==2.7.0
pydantic-settings==2.2.1
aioserial==1.3.1